requirements.txt
Django==4.2.28
google-generativeai==0.3.2
requests==2.31.0
Pillow==10.0.0
python-docx==0.8.11
PyPDF2==3.0.1
